apnaschool/requirements/requirements.txt
Django==3.0.4
django-simple-history==2.8.0